--- app/src/main/java/com/drozdova/danceevents/EventListener.kt
@@ -1,5 +1,0 @@
-package com.drozdova.danceevents
-
-interface EventListener {
-    fun showDetails()
-}
