+++ app/src/main/java/com/drozdova/danceevents/presentation/model/CalendarFragment.kt
@@ -1,15 +1,12 @@
-package com.drozdova.danceevents
+package com.drozdova.danceevents.presentation.model
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.ui.NavigationUI
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
+import com.drozdova.danceevents.R
 import com.drozdova.danceevents.databinding.FragmentCalendarBinding
 
 class CalendarFragment : Fragment(), MonthListener {
