+++ app/src/main/java/com/drozdova/danceevents/presentation/model/EventInfoFragment.kt
@@ -1,4 +1,4 @@
-package com.drozdova.danceevents
+package com.drozdova.danceevents.presentation.model
 
 import android.os.Bundle
 import android.text.method.ScrollingMovementMethod
