+++ app/src/main/java/com/drozdova/danceevents/presentation/model/EventListener.kt
@@ -1,0 +1,5 @@
+package com.drozdova.danceevents.presentation.model
+
+interface EventListener {
+    fun showDetails()
+}
