+++ app/src/main/java/com/drozdova/danceevents/presentation/model/EventsListViewHolder.kt
@@ -1,8 +1,8 @@
-package com.drozdova.danceevents
+package com.drozdova.danceevents.presentation.model
 
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import com.drozdova.danceevents.databinding.ItemEventVerBinding
-import com.drozdova.danceevents.model.EventModel
+import com.drozdova.danceevents.presentation.model.model.EventModel
 
 class EventsListViewHolder(
     private val view: ItemEventVerBinding,
