+++ app/src/main/java/com/drozdova/danceevents/presentation/model/MonthListener.kt
@@ -1,4 +1,4 @@
-package com.drozdova.danceevents
+package com.drozdova.danceevents.presentation.model
 
 interface MonthListener {
     fun showMonthWithEvents()
