+++ app/src/main/java/com/drozdova/danceevents/presentation/model/MonthWithEventsHolder.kt
@@ -1,8 +1,8 @@
-package com.drozdova.danceevents
+package com.drozdova.danceevents.presentation.model
 
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import com.drozdova.danceevents.databinding.ItemEventHorBinding
-import com.drozdova.danceevents.model.EventModel
+import com.drozdova.danceevents.presentation.model.model.EventModel
 
 class MonthWithEventsHolder(
     private val view: ItemEventHorBinding,
