+++ app/src/main/java/com/drozdova/danceevents/presentation/model/SearchFragment.kt
@@ -1,37 +1,36 @@
-package com.drozdova.danceevents
+package com.drozdova.danceevents.presentation.model
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import com.drozdova.danceevents.databinding.FragmentFavEventsBinding
-import com.drozdova.danceevents.model.EventModel
+import com.drozdova.danceevents.R
+import com.drozdova.danceevents.databinding.FragmentSearchBinding
+import com.drozdova.danceevents.presentation.model.model.EventModel
 
-class FavEventsFragment : Fragment(), EventListener {
-    private var _binding: FragmentFavEventsBinding? = null
+class SearchFragment : Fragment(), EventListener {
+    private var _binding: FragmentSearchBinding? = null
     private val binding get() = _binding!!
 
-    private lateinit var adapter: FavEventsAdapter
+    private lateinit var adapter: SearchAdapter
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentFavEventsBinding.inflate(inflater, container, false)
+        _binding = FragmentSearchBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        adapter = FavEventsAdapter(this)
-        binding.rvFavorite.adapter = adapter
+        adapter = SearchAdapter(this)
+        binding.rvSearchEvents.adapter = adapter
 
         val list = listOf(
-            EventModel("Winter Cup 2023", "25.02.2023", "26.02.2023"),
-            EventModel("Child and Youth Week", "25.02.2023", "26.02.2023"),
             EventModel("Minsk Cup", "25.02.2023", "26.02.2023"),
             EventModel("All2TheStep", "25.02.2023", "26.02.2023"),
             EventModel("GolJun", "25.02.2023", "26.02.2023")
@@ -46,6 +45,6 @@ class FavEventsFragment : Fragment(), EventListener {
     }
 
     override fun showDetails() {
-        findNavController().navigate(R.id.action_favEventsFragment_to_eventInfoFragment3)
+        findNavController().navigate(R.id.action_searchFragment_to_eventInfoFragment4)
     }
 }
