+++ app/src/main/res/layout/fragment_calendar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".CalendarFragment"
+    tools:context=".presentation.model.CalendarFragment"
     android:background="@color/black">
 
     <androidx.recyclerview.widget.RecyclerView
