+++ app/src/main/res/layout/fragment_event_info.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EventInfoFragment">
+    tools:context=".presentation.model.EventInfoFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
