+++ app/src/main/res/layout/fragment_events_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".EventsListFragment">
+    tools:context=".presentation.model.EventsListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_events_list"
