+++ app/src/main/res/layout/fragment_fav_events.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".FavEventsFragment">
+    tools:context=".presentation.model.FavEventsFragment">
 
    <androidx.recyclerview.widget.RecyclerView
        android:id="@+id/rv_favorite"
