+++ app/src/main/res/layout/month_item.xml
@@ -7,7 +7,7 @@
     android:layout_height="wrap_content"
     android:orientation="vertical"
     android:layout_margin="3dp"
-    tools:context=".MonthWithEventsFragment">
+    tools:context=".presentation.model.MonthWithEventsFragment">
 
     <com.google.android.material.textview.MaterialTextView
         android:id="@+id/tv_month"
